refactor: move actioncontroller test files to match Rails convention paths
Move 8 test files into controller/ subdirectory to match the Ruby test
directory structure. Rename describe blocks to match Ruby test class
names (FilterTest, RedirectTest, TestCaseTest, etc.).

Import paths updated to reflect new file locations.

This brings actioncontroller from 0/91 to 8/91 matched files.
