+++ packages/actionpack/src/actioncontroller/controller/filters.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
-import { Base } from "./base.js";
-import { Request } from "../actiondispatch/request.js";
-import { Response } from "../actiondispatch/response.js";
+import { Base } from "../base.js";
+import { Request } from "../../actiondispatch/request.js";
+import { Response } from "../../actiondispatch/response.js";
 
 function makeRequest(opts: Record<string, string> = {}): Request {
   return new Request({
@@ -18,7 +18,7 @@ function makeResponse(): Response {
 // ==========================================================================
 // action_controller/filters_test.rb — Controller-level filters
 // ==========================================================================
-describe("ActionController Filters", () => {
+describe("FilterTest", () => {
   it("before_action on controller", async () => {
     const log: string[] = [];
     class AppController extends Base {
