+++ packages/actionpack/src/actioncontroller/controller/rescue.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect } from "vitest";
-import { Base } from "./base.js";
-import { Request } from "../actiondispatch/request.js";
-import { Response } from "../actiondispatch/response.js";
+import { Base } from "../base.js";
+import { Request } from "../../actiondispatch/request.js";
+import { Response } from "../../actiondispatch/response.js";
 
 function makeRequest(opts: Record<string, string> = {}): Request {
   return new Request({
@@ -18,7 +18,7 @@ function makeResponse(): Response {
 // ==========================================================================
 // action_controller/rescue_test.rb
 // ==========================================================================
-describe("ActionController rescue_from", () => {
+describe("RescueControllerTest", () => {
   it("rescues a known error", async () => {
     class AppError extends Error {
       name = "AppError";
